QuantOS/requirements.txt
# Core Math & Data Science
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.35
pandas_ta>=0.3.14b
joblib>=1.3.0
scikit-learn>=1.3.0
scipy>=1.11.0
statsmodels>=0.14.0
ta>=0.11.0
pytz>=2023.3

# Broker Bridges
robin_stocks>=3.2.1
alpaca-py>=0.20.0
ib_insync>=0.9.86
schwab-py>=1.4.0
webull>=0.4.0

# Infrastructure & Database
google-cloud-bigquery>=3.15.0
google-cloud-bigquery-storage>=2.24.0
python-dotenv>=1.0.0
requests>=2.31.0
transformers==4.38.2
torch==2.2.1
colorama>=0.4.6
pyinstaller>=6.0.0
nest_asyncio>=1.5.0
aiohttp>=3.9.0
cryptography>=42.0.0
